change safetab to fit action template and move readme to ROOT with development.md separated out
