+++ README.md
@@ -1,14 +1,26 @@
-# Safetab Action
-Action that can be run by project yaml which gives descriptive 2-way tables 
-of specified variables, and redacts if 5 or less in any cell of the subsequent
-table. 
+### Action Summary
+This is an action that can be called into the 
+`project.yaml`. It requires data in
+a support data file  format (`.csv`, `csv.gz`,
+`.dta`, `feather`), and outputs a folder of markdown files. 
+This action takes in a dataset and a configuration and 
+outputs different types of 2 way tables for descriptive
+statistics. 
 
-### Action Arguments
-The action has one required argument called input_files. This is a file or list 
-of files that you wish to run the safetab action on. 
+Small numbers suppression is applied. A conservative approach
+is adopted where any indiviudal cell that needs to be redacted, 
+results in the whole table being redacted. A log file is created
+that shows what has been redacted. 
 
-##### Config structure
-There is one optional argument `--config`. The config can be a string or a json file.
+### Using Safetab Action 
+The following example blocks should be included 
+in the `project.yaml` file. 
+
+##### Example
+```yaml 
+inputs: 
+  input: tests/test_data/input.feather
+```
 
 The `--config` should contain
 a key called `tables` containing the tables requested. See
@@ -32,22 +44,8 @@ will make this folder if it does not exist.
 
 If you want to change the limit at which redaction occurs, specify this in the 
 `--config` with `redaction_limit`. The default is 5 or below. 
- 
-### Running this action
-You can pip install this package and use as a command line tool. 
-```bash
-safetab [inputfile or list of files] --config [config_json_file or json_str]
 
-# for example to run an input file and config
-safetab data/input.csv --config test_actions_jsons/test1_config.json
-
-# to run list of input files and same config for each file
-safetab data/input.csv data/second_input.csv --config test_actions_jsons/test_json_1.json
-```
-
-### Project yaml
-This action can be invoked from the `project.yaml`. This is passed into json. 
-
+##### Example 
 ```yaml
 actions: 
   safetab_data:
@@ -77,42 +75,3 @@ config:
   output_path: safetab_outputs
   redaction_limit: 5
 ```
-
-## Local Development
-
-For local (non-Docker) development, first install [pyenv][] and execute:
-
-```sh
-pyenv install $(pyenv local)
-```
-
-Then, execute:
-
-```sh
-python3 -m venv venv
-source venv/bin/activate
-pip install --upgrade pip
-pip install -r dev_requirements.txt
-```
-
-## QA
-Run `make` to access Makefile commands. Black, flake8 and mypy are available 
-and have a standard setup. 
-
-## Tests
-
-If you have a local development environment,
-then the following command will write [pytest][]'s output to the terminal:
-
-```sh
-python -m pytest
-```
-
-You can also pass test modules, classes, methods, and functions to pytest:
-
-```sh
-python -m pytest tests/test_processing.py::test_load_study_cohort
-```
-
-[pyenv]: https://github.com/pyenv/pyenv
-[pytest]: https://docs.pytest.org/en/stable/
