+++ action.yaml
@@ -1,0 +1,1 @@
+run: python:latest python action/__main__.py
