+++ action/__main__.py
@@ -2,48 +2,36 @@ import argparse
 import json
 import os
 from pathlib import Path
+from typing import Dict
 
-from safetab.create_tables import output_tables
-from safetab.utils import load_config
-from safetab.version import __version__
-
-
-class ActionConfig:
-    def __init__(self, validator=None):
-        self.validator = validator
-
-    def __call__(self, file_or_string):
-        path = Path(file_or_string)
-        try:
-            if path.exists():
-                with path.open() as f:
-                    config = json.load(f)
-            else:
-                config = json.loads(file_or_string)
-        except json.JSONDecodeError as exc:
-            raise argparse.ArgumentTypeError(f"Could not parse {file_or_string}\n{exc}")
-
-        if self.validator:
-            try:
-                self.validator(config)
-            except Exception as exc:
-                raise argparse.ArgumentTypeError(f"Invalid action config:\n{exc}")
-
-        return config
-
-    @classmethod
-    def add_to_parser(
-        cls, parser, help="The configuration for the safetab action", validator=None
-    ):
-        parser.add_argument(
-            "--config",
-            required=True,
-            help=help,
-            type=ActionConfig(validator),
-        )
+from action.create_tables import output_tables
+from action.utils import load_config
+from action.version import __version__
 
 
-def make_tables(input_files, config):
+def convert_config(file_or_string: str) -> Dict:
+    """
+    Takes in a JSON string or a path to a JSON file and outputs
+    the config as Python object such as a dict
+    Args:
+        file_or_string:
+    Returns:
+        Configuration as loaded JSON
+    """
+    path = Path(file_or_string)
+    try:
+        if path.exists():
+            with path.open() as f:
+                config = json.load(f)
+        else:
+            config = json.loads(file_or_string)
+    except json.JSONDecodeError as exc:
+        raise argparse.ArgumentTypeError(f"Could not parse {file_or_string}\n{exc}")
+
+    return config
+
+
+def run_action(input_files, config):
     """
     :param input_files: The input csvs that the tables are created. These csvs
         must contain the columns names of the table configuration
@@ -65,16 +53,14 @@ def make_tables(input_files, config):
                 - group-by-copd
                 - table-log.txt
     """
-    processed_config = load_config(config)
-
     for input_file in input_files:
         input_filename_with_ext = os.path.basename(input_file)
         input_filename = os.path.splitext(input_filename_with_ext)[0]
         output_tables(
             data_csv=input_file,
-            table_config=processed_config["tables"],
-            output_dir=f"{processed_config['output_path']}/{input_filename}_tables",
-            limit=processed_config["redaction_limit"],
+            table_config=config["tables"],
+            output_dir=f"{config['output_path']}/{input_filename}_tables",
+            limit=config["redaction_limit"],
         )
 
 
@@ -91,21 +77,28 @@ def main():
     )
 
     # configurations
-    ActionConfig.add_to_parser(parser)
+    parser.add_argument(
+        "--config",
+        help="Configuration of the action - either a JSON str, or a path to a JSON file",
+    )
 
     # version
-    parser.add_argument("--version", action="version", version=f"safetab {__version__}")
+    parser.add_argument("--version", action="version", version=f"action {__version__}")
 
     # input files
     parser.add_argument(
-        "input_files", nargs="*", help="Files that safetab will be run on"
+        "input_files", nargs="*", help="Files that action will be run on"
     )
 
     # parse args
     args = parser.parse_args()
 
+    # convert config path to config dict
+    config_dict = convert_config(args.config)
+    processed_config = load_config(config_dict)
+
     # pass the data from the json file to the output_tables arguments
-    make_tables(input_files=args.input_files, config=args.config)
+    run_action(input_files=args.input_files, config=processed_config)
 
 
 if __name__ == "__main__":
