+++ action/utils.py
@@ -5,12 +5,12 @@ DEFAULTS = {"output_path": "safetab_tables", "limit": 5}
 
 def load_config(config) -> Dict:
     """
-    Takes in safetab configuration and changes these key-value pairs
-    where indicated by the safetab config. All other key-value pairs
+    Takes in action configuration and changes these key-value pairs
+    where indicated by the action config. All other key-value pairs
     are left as default values
 
     Args:
-        config (dict): dictionary of the safetab configuration
+        config (dict): dictionary of the action configuration
             taken from json
 
     Returns:
