+++ test_table_outputs/death_2_way_tabs/ageband vs death.md
@@ -1,0 +1,27 @@
+ageband vs death
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   1 |   Total |
+|:----------|----:|----:|--------:|
+| 60 - 69   |  13 |  11 |      24 |
+| 70 - 79   |  11 |  25 |      36 |
+| Total     |  24 |  36 |      60 |
+
+ageband vs death with Column Percentages
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   Column1 % |   1 |   Column2 % |
+|:----------|----:|------------:|----:|------------:|
+| 60 - 69   |  13 |        54.2 |  11 |        30.6 |
+| 70 - 79   |  11 |        45.8 |  25 |        69.4 |
+
+ageband vs death with Row Percentages
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   Row1 % |   1 |   Row2 % |
+|:----------|----:|---------:|----:|---------:|
+| 60 - 69   |  13 |     54.2 |  11 |     45.8 |
+| 70 - 79   |  11 |     30.6 |  25 |     69.4 |
