+++ test_table_outputs/death_2_way_tabs/copd vs death.md
@@ -1,0 +1,27 @@
+copd vs death
+
+ROWS: copd, COLUMNS: death
+
+| copd   |   0 |   1 |   Total |
+|:-------|----:|----:|--------:|
+| 0      |  14 |  22 |      36 |
+| 1      |  10 |  14 |      24 |
+| Total  |  24 |  36 |      60 |
+
+copd vs death with Column Percentages
+
+ROWS: copd, COLUMNS: death
+
+|   copd |   0 |   Column1 % |   1 |   Column2 % |
+|-------:|----:|------------:|----:|------------:|
+|      0 |  14 |        58.3 |  22 |        61.1 |
+|      1 |  10 |        41.7 |  14 |        38.9 |
+
+copd vs death with Row Percentages
+
+ROWS: copd, COLUMNS: death
+
+|   copd |   0 |   Row1 % |   1 |   Row2 % |
+|-------:|----:|---------:|----:|---------:|
+|      0 |  14 |     38.9 |  22 |     61.1 |
+|      1 |  10 |     41.7 |  14 |     58.3 |
