+++ test_table_outputs/death_2_way_tabs/sex vs death.md
@@ -1,0 +1,27 @@
+sex vs death
+
+ROWS: sex, COLUMNS: death
+
+| sex   |   0 |   1 |   Total |
+|:------|----:|----:|--------:|
+| F     |  16 |  16 |      32 |
+| M     |   8 |  20 |      28 |
+| Total |  24 |  36 |      60 |
+
+sex vs death with Column Percentages
+
+ROWS: sex, COLUMNS: death
+
+| sex   |   0 |   Column1 % |   1 |   Column2 % |
+|:------|----:|------------:|----:|------------:|
+| F     |  16 |        66.7 |  16 |        44.4 |
+| M     |   8 |        33.3 |  20 |        55.6 |
+
+sex vs death with Row Percentages
+
+ROWS: sex, COLUMNS: death
+
+| sex   |   0 |   Row1 % |   1 |   Row2 % |
+|:------|----:|---------:|----:|---------:|
+| F     |  16 |     50   |  16 |     50   |
+| M     |   8 |     28.6 |  20 |     71.4 |
