+++ test_table_outputs/grouped_by_sex/df_F - ageband vs death.md
@@ -1,0 +1,27 @@
+ageband vs death
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   1 |   Total |
+|:----------|----:|----:|--------:|
+| 60 - 69   |  10 |  10 |      20 |
+| 70 - 79   |   6 |   6 |      12 |
+| Total     |  16 |  16 |      32 |
+
+ageband vs death with Column Percentages
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   Column1 % |   1 |   Column2 % |
+|:----------|----:|------------:|----:|------------:|
+| 60 - 69   |  10 |        62.5 |  10 |        62.5 |
+| 70 - 79   |   6 |        37.5 |   6 |        37.5 |
+
+ageband vs death with Row Percentages
+
+ROWS: ageband, COLUMNS: death
+
+| ageband   |   0 |   Row1 % |   1 |   Row2 % |
+|:----------|----:|---------:|----:|---------:|
+| 60 - 69   |  10 |       50 |  10 |       50 |
+| 70 - 79   |   6 |       50 |   6 |       50 |
