+++ test_table_outputs/grouped_by_sex/df_F - copd vs death.md
@@ -1,0 +1,27 @@
+copd vs death
+
+ROWS: copd, COLUMNS: death
+
+| copd   |   0 |   1 |   Total |
+|:-------|----:|----:|--------:|
+| 0      |  10 |   6 |      16 |
+| 1      |   6 |  10 |      16 |
+| Total  |  16 |  16 |      32 |
+
+copd vs death with Column Percentages
+
+ROWS: copd, COLUMNS: death
+
+|   copd |   0 |   Column1 % |   1 |   Column2 % |
+|-------:|----:|------------:|----:|------------:|
+|      0 |  10 |        62.5 |   6 |        37.5 |
+|      1 |   6 |        37.5 |  10 |        62.5 |
+
+copd vs death with Row Percentages
+
+ROWS: copd, COLUMNS: death
+
+|   copd |   0 |   Row1 % |   1 |   Row2 % |
+|-------:|----:|---------:|----:|---------:|
+|      0 |  10 |     62.5 |   6 |     37.5 |
+|      1 |   6 |     37.5 |  10 |     62.5 |
