+++ test_table_outputs/simple_2_way_tabs/ageband vs copd.md
@@ -1,0 +1,27 @@
+ageband vs copd
+
+ROWS: ageband, COLUMNS: copd
+
+| ageband   |   0 |   1 |   Total |
+|:----------|----:|----:|--------:|
+| 60 - 69   |  11 |  13 |      24 |
+| 70 - 79   |  25 |  11 |      36 |
+| Total     |  36 |  24 |      60 |
+
+ageband vs copd with Column Percentages
+
+ROWS: ageband, COLUMNS: copd
+
+| ageband   |   0 |   Column1 % |   1 |   Column2 % |
+|:----------|----:|------------:|----:|------------:|
+| 60 - 69   |  11 |        30.6 |  13 |        54.2 |
+| 70 - 79   |  25 |        69.4 |  11 |        45.8 |
+
+ageband vs copd with Row Percentages
+
+ROWS: ageband, COLUMNS: copd
+
+| ageband   |   0 |   Row1 % |   1 |   Row2 % |
+|:----------|----:|---------:|----:|---------:|
+| 60 - 69   |  11 |     45.8 |  13 |     54.2 |
+| 70 - 79   |  25 |     69.4 |  11 |     30.6 |
