+++ test_table_outputs/simple_2_way_tabs/sex vs copd.md
@@ -1,0 +1,27 @@
+sex vs copd
+
+ROWS: sex, COLUMNS: copd
+
+| sex   |   0 |   1 |   Total |
+|:------|----:|----:|--------:|
+| F     |  16 |  16 |      32 |
+| M     |  20 |   8 |      28 |
+| Total |  36 |  24 |      60 |
+
+sex vs copd with Column Percentages
+
+ROWS: sex, COLUMNS: copd
+
+| sex   |   0 |   Column1 % |   1 |   Column2 % |
+|:------|----:|------------:|----:|------------:|
+| F     |  16 |        44.4 |  16 |        66.7 |
+| M     |  20 |        55.6 |   8 |        33.3 |
+
+sex vs copd with Row Percentages
+
+ROWS: sex, COLUMNS: copd
+
+| sex   |   0 |   Row1 % |   1 |   Row2 % |
+|:------|----:|---------:|----:|---------:|
+| F     |  16 |     50   |  16 |     50   |
+| M     |  20 |     71.4 |   8 |     28.6 |
