+++ test_table_outputs/table_log.txt
@@ -1,0 +1,12 @@
+simple_2_way_tabs - sex vs ageband - Table REDACTED, small numbers
+simple_2_way_tabs - sex vs copd - Table created
+simple_2_way_tabs - ageband vs copd - Table created
+death_2_way_tabs - sex vs death - Table created
+death_2_way_tabs - ageband vs death - Table created
+death_2_way_tabs - copd vs death - Table created
+grouped_by_sex - df_F - ageband vs copd - Table REDACTED, small numbers
+grouped_by_sex - df_F - ageband vs death - Table created
+grouped_by_sex - df_F - copd vs death - Table created
+grouped_by_sex - df_M - ageband vs copd - Table REDACTED, small numbers
+grouped_by_sex - df_M - ageband vs death - Table REDACTED, small numbers
+grouped_by_sex - df_M - copd vs death - Table REDACTED, small numbers
