+++ tests/test_find_and_save_tools.py
@@ -1,7 +1,7 @@
 import pytest
 
-from safetab.errors import ImportActionError
-from safetab.find_save_tools import import_data
+from action.errors import ImportActionError
+from action.find_save_tools import import_data
 
 TEST_DATA_CSV = "tests/test_data/test_data.csv"
